# openapi/account.oas.yaml
openapi: 3.0.3
info:
  title: Account Service
  version: 1.0.0
servers:
  - url: https://{apiId}.execute-api.eu-north-1.amazonaws.com
    variables:
      apiId:
        default: 6wz68m95ij
paths:
  /accounts:
    get:
      operationId: listAccounts
      summary: List all accounts
      parameters:
        - name: customerId
          in: query
          required: false
          schema:
            type: string
            format: uuid
      responses:
        '200':
          description: List of accounts
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/Account'
    post:
      operationId: createAccount
      summary: Create an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CreateAccount'
      responses:
        '201':
          description: Account created
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          $ref: '#/components/responses/BadRequest'
  /accounts/{id}:
    parameters:
      - name: id
        in: path
        required: true
        schema:
          type: string
          format: uuid
    get:
      operationId: getAccount
      summary: Get an account by ID
      responses:
        '200':
          description: Account found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    patch:
      operationId: updateAccount
      summary: Update an account
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/UpdateAccount'
      responses:
        '200':
          description: Updated
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Account'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
    delete:
      operationId: deleteAccount
      summary: Delete an account by ID
      responses:
        '204':
          description: Account deleted
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
components:
  schemas:
    Account:
      type: object
      required:
        - id
        - name
        - iban
        - currency
        - customerId
      properties:
        id:
          type: string
          format: uuid
        name:
          type: string
        iban:
          type: string
          pattern: '^[A-Z]{2}[0-9]{2}[A-Z0-9]{1,30}$'
          minLength: 15
          maxLength: 34
          example: DK5000400440116243
        currency:
          type: string
          description: ISO 4217 currency code
          pattern: '^[A-Z]{3}$'
          minLength: 3
          maxLength: 3
          example: DKK
        customerId:
          type: string
          format: uuid
    UpdateAccount:
      type: object
      required: [name]
      properties:
        name:
          type: string

    CreateAccount:
      type: object
      required:
        - name
        - iban
        - currency
        - customerId
      properties:
        name:
          type: string
        iban:
          type: string
          pattern: '^[A-Z]{2}[0-9]{2}[A-Z0-9]{1,30}$'
          minLength: 15
          maxLength: 34
          example: DK5000400440116243
        currency:
          type: string
          description: ISO 4217 currency code
          pattern: '^[A-Z]{3}$'
          minLength: 3
          maxLength: 3
          example: DKK
        customerId:
          type: string
          format: uuid
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            type: object
            required:
              - error
            properties:
              error:
                type: string
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            type: object
            required:
              - error
            properties:
              error:
                type: string